+++ demo/test/components/previews/primer/docs/box_component_preview.rb
@@ -1,9 +1,0 @@
-module Primer
-  module Docs
-    class BoxComponentPreview < ViewComponent::Preview
-      def default; end
-
-      def color_and_padding; end
-    end
-  end
-end