+++ demo/test/components/previews/primer/docs/breadcrumb_component_preview.rb
@@ -1,7 +1,0 @@
-module Primer
-  module Docs
-    class BreadcrumbComponentPreview < ViewComponent::Preview
-      def basic; end
-    end
-  end
-end