+++ demo/test/components/previews/primer/docs/layout_component_preview.rb
@@ -1,9 +1,0 @@
-module Primer
-  module Docs
-    class LayoutComponentPreview < ViewComponent::Preview
-      def default; end
-
-      def left_sidebar; end
-    end
-  end
-end